Module aws_sdk_dlm::model

Modules

action

See Action

create_rule

See CreateRule

cross_region_copy_action

See CrossRegionCopyAction

cross_region_copy_deprecate_rule

See CrossRegionCopyDeprecateRule

cross_region_copy_retain_rule

See CrossRegionCopyRetainRule

cross_region_copy_rule

See CrossRegionCopyRule

deprecate_rule

See DeprecateRule

encryption_configuration

See EncryptionConfiguration

event_parameters

See EventParameters

event_source

See EventSource

fast_restore_rule

See FastRestoreRule

lifecycle_policy

See LifecyclePolicy

lifecycle_policy_summary

See LifecyclePolicySummary

parameters

See Parameters

policy_details

See PolicyDetails

retain_rule

See RetainRule

schedule

See Schedule

share_rule

See ShareRule

tag

See Tag

Structs

Action

Specifies an action for an event-based policy.

CreateRule

Specifies when to create snapshots of EBS volumes.

CrossRegionCopyAction

Specifies a rule for copying shared snapshots across Regions.

CrossRegionCopyDeprecateRule

Specifies an AMI deprecation rule for cross-Region AMI copies created by a cross-Region copy rule.

CrossRegionCopyRetainRule

Specifies the retention rule for cross-Region snapshot copies.

CrossRegionCopyRule

Specifies a rule for cross-Region snapshot copies.

DeprecateRule

Specifies an AMI deprecation rule for a schedule.

EncryptionConfiguration

Specifies the encryption settings for shared snapshots that are copied across Regions.

EventParameters

Specifies an event that triggers an event-based policy.

EventSource

Specifies an event that triggers an event-based policy.

FastRestoreRule

Specifies a rule for enabling fast snapshot restore. You can enable fast snapshot

LifecyclePolicy

Detailed information about a lifecycle policy.

LifecyclePolicySummary

Summary information about a lifecycle policy.

Parameters

Specifies optional parameters to add to a policy. The set of valid parameters depends

PolicyDetails

Specifies the configuration of a lifecycle policy.

RetainRule

Specifies the retention rule for a lifecycle policy. You can retain snapshots based on

Schedule

Specifies a backup schedule for a snapshot or AMI lifecycle policy.

ShareRule

Specifies a rule for sharing snapshots across Amazon Web Services accounts.

Tag

Specifies a tag for a resource.

Enums

EventSourceValues

EventTypeValues

GettablePolicyStateValues

IntervalUnitValues

LocationValues

PolicyTypeValues

ResourceLocationValues

ResourceTypeValues

RetentionIntervalUnitValues

SettablePolicyStateValues